pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "podflow"
version = "0.1.0"
description = "End-to-end podcast automation pipeline"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "ffmpeg-python>=0.2.0",
    "mutagen>=1.47",
    "openai>=1.0",
    "anthropic>=0.20",
    "google-api-python-client>=2.100",
    "google-auth-oauthlib>=1.1",
    "feedgen>=1.0",
    "boto3>=1.28",
    "rich>=13.0",
]

[project.optional-dependencies]
whisper-local = ["openai-whisper>=20230918"]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.12",
]

[project.scripts]
podflow = "podflow.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]
